+++ src/analysis/pitch_extraction_by_world.cc
@@ -90,14 +90,14 @@ bool PitchExtractionByWorld::Get(
     option.allowed_range = voicing_threshold_;
 
     const int tmp_length(
-        world::GetSamplesForDIO(sampling_rate_, waveform.size(), frame_period));
+        world::GetSamplesForDIO(static_cast<int>(sampling_rate_), static_cast<int>(waveform.size()), frame_period));
     std::vector<double> time_axis(tmp_length);
     std::vector<double> tmp_f0(tmp_length);
-    world::Dio(&(waveform[0]), waveform.size(), sampling_rate_, &option,
+    world::Dio(&(waveform[0]), static_cast<int>(waveform.size()), static_cast<int>(sampling_rate_), &option,
                &(time_axis[0]), &(tmp_f0[0]));
 
-    const int target_length(
-        std::ceil(static_cast<double>(waveform.size()) / frame_shift_));
+    const int target_length(static_cast<int>(
+        std::ceil(static_cast<double>(waveform.size()) / frame_shift_)));
     if (target_length < tmp_length) {
       tmp_f0.resize(target_length);
     }
