+++ src/compression/uniform_quantization.cc
@@ -87,12 +87,12 @@ bool UniformQuantization::Run(double input, int* output) const {
   int index;
   switch (quantization_type_) {
     case kMidRise: {
-      index = std::floor(input * inverse_step_size_) + quantization_levels_ / 2;
+      index = static_cast<int>(std::floor(input * inverse_step_size_) + quantization_levels_ / 2);
       break;
     }
     case kMidTread: {
-      index = std::round(input * inverse_step_size_) +
-              (quantization_levels_ - 1) / 2;
+      index = static_cast<int>(std::round(input * inverse_step_size_) +
+              (quantization_levels_ - 1) / 2);
       break;
     }
     default: { return false; }
